% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_AlternatedLFMM.R
\name{AlternatedLFMMMethod}
\alias{AlternatedLFMMMethod}
\title{||G  - U V^t - X B || + alpha ||U||^2 + beta ||V B^t||^2}
\usage{
AlternatedLFMMMethod(K, hypothesis.testing.method = Zscore(NormalZscore(),
  AnalyticSigma2Functor()), it.max = 100, err.max = 1e-06, alpha = 10,
  beta = 10, center = TRUE, name = "AlternatedLFMMMethod")
}
\description{
||G  - U V^t - X B || + alpha ||U||^2 + beta ||V B^t||^2
}

