% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_AlternatedSvdLFMMMethod.R
\name{AlternatedSvdLFMMMethod}
\alias{AlternatedSvdLFMMMethod}
\title{||G  - U V^t - X B || + lambda ||B||^2}
\usage{
AlternatedSvdLFMMMethod(K, it.max = 100, err.max = 1e-06,
  hypothesis.testing.method = Zscore(NormalZscore(), AnalyticSigma2Functor()),
  lambda = 0, center = TRUE, name = "AlternatedSvdLFMMMethod",
  nickname = NULL)
}
\description{
||G  - U V^t - X B || + lambda ||B||^2
}

