% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_zscore.R
\name{AnalyticSigma2Functor}
\alias{AnalyticSigma2Functor}
\title{analytic Sigma2 functor}
\usage{
AnalyticSigma2Functor()
}
\description{
B.Sigma2 was computed during the fit step.
}

