% Generated by roxygen2: do not edit by hand
% Please edit documentation in 3Article/3Article_Celiac.R
\name{Article3_Celiac_fake_sampler}
\alias{Article3_Celiac_fake_sampler}
\title{Just to work with MethodBatchExperiment}
\usage{
Article3_Celiac_fake_sampler()
}
\description{
Just to work with MethodBatchExperiment
}

