% Generated by roxygen2: do not edit by hand
% Please edit documentation in 3Article/3Article_MethodComparison.R
\name{Article3_MethodComparison_plot_GIF}
\alias{Article3_MethodComparison_plot_GIF}
\title{This function compute GIF on null loci, so the GIF must be equal to 1}
\usage{
Article3_MethodComparison_plot_GIF(exp)
}
\description{
We use use a quantile function to compute score2 from pvalue... ;)
}

