% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_refractor.R
\name{associations_test}
\alias{associations_test}
\title{code found here: https://github.com/cozygene/refactor ... surement codé par un expert en R ^_^'}
\usage{
associations_test(O, y, model_append)
}
\description{
code found here: https://github.com/cozygene/refactor ... surement codé par un expert en R ^_^'
}

