% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_bootstrap.R
\name{Bootstrap}
\alias{Bootstrap}
\title{normal_zscore functor with boostrap}
\usage{
Bootstrap(bootstrap.func, bootstrap.rep)
}
\description{
B.Sigma2 are computed with bootstrap.func.
We assume that zscore~N(0,1) under H0
}

