% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_LinearMethod.R
\name{ClassicLinearMethod}
\alias{ClassicLinearMethod}
\title{G = X B + E}
\usage{
ClassicLinearMethod(center = TRUE, hypothesis.testing.method = Zscore(),
  nickname = NULL)
}
\description{
G = X B + E
}

