% Generated by roxygen2: do not edit by hand
% Please edit documentation in Experiment/Experiment_ComparisonExperiment.R
\name{ComparisonExperiment}
\alias{ComparisonExperiment}
\title{ComparisonExperiment class}
\usage{
ComparisonExperiment(sampler, ...,
  description = "comparison of algorithm on a sampler")
}
\description{
ComparisonExperiment class
}

