% Generated by roxygen2: do not edit by hand
% Please edit documentation in Experiment/Experiment_Helpers.R
\name{ComparisonExperiment_GridRun}
\alias{ComparisonExperiment_GridRun}
\title{run ComparisonExperiment on a grid}
\usage{
ComparisonExperiment_GridRun(method.constructor, s, ms, Ks, lambdas, gif, save)
}
\description{
run ComparisonExperiment on a grid
}

