% Generated by roxygen2: do not edit by hand
% Please edit documentation in Helpers/helpers.R
\name{CRANDep}
\alias{CRANDep}
\title{Install CRAN dep}
\usage{
CRANDep(pkgs = c("ggplot2", "cowplot", "tidyr", "tibble", "purrr", "RSQLite",
  "magrittr", "MASS", "foreach", "tools", "readr", "knitr", "gtools",
  "doParallel", "parallel", "FAMT", "assertthat", "fdrtool", "crayon"))
}
\description{
Install CRAN dep
}

