% Generated by roxygen2: do not edit by hand
% Please edit documentation in CrossValidation/CrossValidation_rowwise.R
\name{CrossValidation_rowwise}
\alias{CrossValidation_rowwise}
\title{Cross validation on row}
\usage{
CrossValidation_rowwise(m, dat, row.left.out.func, col.left.out.func, lambdas,
  Ks)
}
\description{
Cross validation on row
}

