% Generated by roxygen2: do not edit by hand
% Please edit documentation in Experiment/Experiment_CrossValidation.R
\name{CrossValidationExperiment}
\alias{CrossValidationExperiment}
\title{An experiment to run crossvalidation on a method}
\usage{
CrossValidationExperiment(m, s, lambdas, kfold,
  crossvalidation.func = crossvalidation_kfold_indiv,
  description = paste("Cross validation of ", s$nickname, "by", m$nickname,
  "for lambda =", paste(lambdas, collapse = "|")))
}
\description{
An experiment to run crossvalidation on a method
}

