% Generated by roxygen2: do not edit by hand
% Please edit documentation in Experiment/Experiment_DebugExperiment.R
\name{DebugExperiment}
\alias{DebugExperiment}
\title{An experiment to run a method on several sampler and see results}
\usage{
DebugExperiment(method, sampler.list, parameter.list)
}
\description{
An experiment to run a method on several sampler and see results
}

