% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_LinearMethod.R
\name{epsilon.sigma2}
\alias{epsilon.sigma2}
\title{formula find in Cule et al. 2011}
\usage{
epsilon.sigma2(epsilon, reduced.df)
}
\arguments{
\item{reduced.df}{use to compute effective.degree.freedom}
}
\description{
formula find in Cule et al. 2011
}

