% Generated by roxygen2: do not edit by hand
% Please edit documentation in Helpers/FDRControlHelpers.R
\name{ExpectedFDR_trueFDR_Power}
\alias{ExpectedFDR_trueFDR_Power}
\title{compute expeted true FDR and power}
\usage{
ExpectedFDR_trueFDR_Power(pvalue, outlier)
}
\arguments{
\item{pvalue}{a vector of pvalue for each snips}

\item{outlier}{a vector of outlier snps index}
}
\description{
assumig we know m0 the number not outlier. We use
P(H0|positif) = P(positif | H0) * P(H0) / P(positif)
}

