% Generated by roxygen2: do not edit by hand
% Please edit documentation in Experiment/Experiment.R
\name{Experiment}
\alias{Experiment}
\title{Experiment class}
\usage{
Experiment(name = "Default", date = as.character(Sys.time()),
  seed = sample.int(.Machine$integer.max, 1),
  machine = as.character(Sys.info()[4]), runtime = 0, dumpfile = "NULL",
  description = "NULL", benchmakdir = getOption("thesis.dump.file"))
}
\description{
Experiment class
}

