% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_FAMT.R
\name{FAMTMethod}
\alias{FAMTMethod}
\title{FAMT package wrapper}
\usage{
FAMTMethod(K, hypothesis.testing.method = NULL, name = "FAMTMethod",
  nickname = NULL)
}
\description{
FAMT package wrapper
}

