% Generated by roxygen2: do not edit by hand
% Please edit documentation in Experiment/Experiment_FDRControlExperiment.R
\name{FDRControlExperiment}
\alias{FDRControlExperiment}
\title{FDRControlExperiment}
\usage{
FDRControlExperiment(nb.rep, sampler, ...,
  seed = sample.int(.Machine$integer.max, 1))
}
\description{
FDRControlExperiment
}

