% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_zscore.R
\name{FdrtoolsCalibratedZscore}
\alias{FdrtoolsCalibratedZscore}
\title{pvalue computed with fdrtool package}
\usage{
FdrtoolsCalibratedZscore()
}
\description{
pvalue computed with fdrtool package
}

