% Generated by roxygen2: do not edit by hand
% Please edit documentation in Helpers/Helpers_logs.R
\name{flog.debug}
\alias{flog.debug}
\title{futile.logger wrapper}
\usage{
flog.debug(msg, ..., name = "ThesisRpackage", capture = FALSE)
}
\arguments{
\item{msg}{The message}
}
\description{
futile.logger wrapper
}

