% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_FromTrueData.R
\name{FromTrueSampler}
\alias{FromTrueSampler}
\title{see my notebook (18/01/2017), sample a generative model from
true dataset.}
\usage{
FromTrueSampler(G.file, n, L, K, prop.outlier, rho = NULL, cs = NULL,
  round = FALSE, sd.V.rho = 1, rho.E = 1,
  B.outlier.sampler = function(n, mean, sd) rnorm(n, mean, sd))
}
\arguments{
\item{rho}{must be between 0 and 1. the proportion of variance explained by X.
If NULL just a sum between UV^T and XB}
}
\description{
see my notebook (18/01/2017), sample a generative model from
true dataset.
}

