% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_FromTrueData2.R
\name{FromTrueSampler2}
\alias{FromTrueSampler2}
\title{see my notebook (13/06/2017), sample a generative model from
true dataset.}
\usage{
FromTrueSampler2(G.file, K, prop.outlier, cs = NULL, pca.file = NULL,
  rho.B = 1, n = NULL, L = NULL)
}
\arguments{
\item{cs}{if NULL sample. Must be of size K.}
}
\description{
see my notebook (13/06/2017), sample a generative model from
true dataset.
}

