% Generated by roxygen2: do not edit by hand
% Please edit documentation in DataSet/DataSet_GenerativeDataSet.R
\name{GenerativeDataSet}
\alias{GenerativeDataSet}
\title{GenerativeDataSet class}
\usage{
GenerativeDataSet(n = 1, L = 10, G = matrix(0, 1, L), X = matrix(0, n,
  1), U = matrix(0, n, 1), V = matrix(0, L, 1), B = matrix(0, 1, L),
  epsilon = matrix(0, n, L), outlier = c(), mu = matrix(0, 1, L))
}
\description{
GenerativeDataSet class
}

