% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_zscore.R
\name{GifCalibratedZscore}
\alias{GifCalibratedZscore}
\title{GifCalibratedZscore functor}
\usage{
GifCalibratedZscore()
}
\description{
We assume that zscore~N(0,sd) under H0 so (find ref...)
z^2 / gif  ~ chi2(df = 1) with gif = median(z^2) / qchist(0.5)
}

