% Generated by roxygen2: do not edit by hand
% Please edit documentation in Helpers/Helpers_plots.R
\name{gplot_stat}
\alias{gplot_stat}
\title{Plot stat}
\usage{
gplot_stat(..., outlier)
}
\description{
Plot stat
}
\examples{
library(Article3Package)

K <- 5
s <- NormalSampler(100, 1000, K,
                   prop.outlier = 0.02,
                   c = 0.8) \%>\%
  MissingValueSampler(missing.prop = 0.0)

m <- ClassicLinearMethod()
dat <- sampl(s)
m <- run(m, dat)
gplot_stat(m$score[1,] outlier = dat$outlier) +
 geom_point(aes(x = index, y = stat, color = outlier)) +

}

