% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_LassoLFMMMethod.R
\name{Lasso_LambdaGammaRange}
\alias{Lasso_LambdaGammaRange}
\title{compute a range of lambda}
\usage{
Lasso_LambdaGammaRange(dat, lambda.K, Ks)
}
\description{
compute a range of lambda
}

