% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_LassoLFMMMethod.R
\name{Lasso_LambdaRange}
\alias{Lasso_LambdaRange}
\title{compute a range of gamma and lambda}
\usage{
Lasso_LambdaRange(dat, K, lambda.K = 100, lambda.eps = 0.001)
}
\description{
compute a range of gamma and lambda
}

