% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_LassoLFMMMethod.R
\name{LassoLFMMMethod}
\alias{LassoLFMMMethod}
\title{||G  - U V^t - X B || + lambda |B| + gamma |UV^T|}
\usage{
LassoLFMMMethod(K, it.max, err.max, gamma = NULL, lambda = 0,
  lambda.K = 100, lambda.eps = 0.001, sparse.prop = NULL, center = TRUE,
  soft = TRUE, name = "LassoLFMMMethod", nickname = "LassoLFMMMethod",
  hypothesis.testing.method = NULL)
}
\description{
||G  - U V^t - X B || + lambda |B| + gamma |UV^T|
}

