% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_lm.R
\name{lm_zscore}
\alias{lm_zscore}
\title{lm + zscore on G - C}
\usage{
lm_zscore(calibrate = FALSE, correctionByC = TRUE, center = TRUE,
  sigma.computation = c("lm", "bootstrap", "lm+df"),
  bootstrap.func = PairedBoostrap, bootstrap.rep = 30)
}
\description{
perform a linear reg on G - C then make a classic zscore hypothesis testing
}

