% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_GenerativeSampler.R
\name{LogisticSampler}
\alias{LogisticSampler}
\title{Title}
\usage{
LogisticSampler(n, L, K, ploidy = 1, prop.outlier = 0.02, sigma = 0,
  c = 0.6, mean.B = 0, sd.B = 1, sd.mu = 1, mean.mu = 0.5, sd.U = 1,
  sd.V = 1)
}
\description{
Title
}

