% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_LinearMethod.R
\name{loglr.ClassicLinearMethod}
\alias{loglr.ClassicLinearMethod}
\title{loglr of classic lineara method}
\usage{
\method{loglr}{ClassicLinearMethod}(m, dat)
}
\description{
for linear regresion the log likelyhood ratio is
 - 1 / (2 * sigma^2) (RSS_0 - RSS) whit murphy's book notation
}

