% Generated by roxygen2: do not edit by hand
% Please edit documentation in Helpers/helpers.R
\name{long_init}
\alias{long_init}
\title{Use it before running the core of a long function}
\usage{
long_init(cluster.nb, bypass, log.file = NULL)
}
\description{
Use it before running the core of a long function
}

