% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_zscore.R
\name{MadCalibratedZscore}
\alias{MadCalibratedZscore}
\title{MadCalibratedZscore functor}
\usage{
MadCalibratedZscore()
}
\description{
We assume that zscore~N(0,sd) under H0 and try to estimate sd with the mad
}

