% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_MafFilter.R
\name{MafFilterSampler}
\alias{MafFilterSampler}
\title{run the sampler and filter minor allele frequency}
\usage{
MafFilterSampler(sampler, maf.threshold)
}
\description{
run the sampler and filter minor allele frequency
}

