% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod.R
\name{Method}
\alias{Method}
\title{Sampler class}
\usage{
Method(name = "A method", hypothesis.testing.method = NULL,
  impute.genotype.method = imputeByMean(), nickname = NULL)
}
\description{
Sampler class
}

