% Generated by roxygen2: do not edit by hand
% Please edit documentation in Experiment/Experiment_MethodBatchExperiment.R
\name{MethodBatchExperiment}
\alias{MethodBatchExperiment}
\title{Run a batch of method on a dataset}
\usage{
MethodBatchExperiment(data.name, s, method.batch, cluster.nb = NULL)
}
\description{
Run a batch of method on a dataset
}

