% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_GenerativeSampler.R
\name{NormalSampler}
\alias{NormalSampler}
\title{Title}
\usage{
NormalSampler(n, L, K, prop.outlier = 0.02, sigma = 0.2, c = 0.6,
  cs = NULL, mean.B = 0, sd.B = 1, sd.mu = 1, mean.mu = 0.5,
  sd.U = 1, sd.V = 1)
}
\description{
Title
}

