% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_GenerativeSampler.R
\name{NormalSampler2}
\alias{NormalSampler2}
\title{see my labnotebook (2/02/2017)}
\usage{
NormalSampler2(n, L, K, prop.outlier = 0.02, sigma = 0.2, cs = c(0.6,
  rep(0, K - 1)), mean.B = 0, sd.B = 1, sd.mu = 1, mean.mu = 0.5,
  sd.U = 1, sd.V = 1)
}
\description{
see my labnotebook (2/02/2017)
}

