% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_NuclearLFMM.R
\name{NuclearLFMMMethod}
\alias{NuclearLFMMMethod}
\title{||G  - C - X B || + gamma ||C||_*}
\usage{
NuclearLFMMMethod(K, hypothesis.testing.method = Zscore(NormalZscore(),
  AnalyticSigma2Functor()), it.max = 100, err.max = 1e-06, gamma = NULL,
  lambda = 0, lambda.K = 100, lambda.eps = 0.001, sparse.prop = NULL,
  center = TRUE, name = "NuclearLFMMMethod", lasso = FALSE, soft = TRUE,
  nickname = NULL)
}
\arguments{
\item{gamma}{if -1 try to compute a gamma such that rk(C) = K.}
}
\description{
||G  - C - X B || + gamma ||C||_*
}

