% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod.R
\name{numLatentVarEstimation}
\alias{numLatentVarEstimation}
\title{Estimated the number of latent variable}
\usage{
numLatentVarEstimation(m, dat)
}
\description{
Estimated the number of latent variable
}

