% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_bootstrap.R
\name{PairedBoostrap}
\alias{PairedBoostrap}
\title{boostrap method described in 8.3 of http://www.stat.ncsu.edu/people/lu/courses/ST505/Ch8.pdf}
\usage{
PairedBoostrap(dat, m)
}
\description{
boostrap method described in 8.3 of http://www.stat.ncsu.edu/people/lu/courses/ST505/Ch8.pdf
}

