% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_LinearMethod.R
\name{PCAClassicLinearMethod}
\alias{PCAClassicLinearMethod}
\title{PCA + linear method}
\usage{
PCAClassicLinearMethod(K, center = TRUE,
  hypothesis.testing.method = Zscore(), nickname = NULL,
  assumingStructure = FALSE)
}
\description{
svd(G) -> U ...
G = (U|X) B + E
}

