% Generated by roxygen2: do not edit by hand
% Please edit documentation in Experiment/Experiment_PCA.R
\name{PCAExperiment}
\alias{PCAExperiment}
\title{An experiment to run PCA on a data set.}
\usage{
PCAExperiment(s, description = "Run PCA on dataset",
  seed = sample.int(.Machine$integer.max, 1))
}
\description{
Then compute correlation between U and X.
}

