% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_perm.R
\name{PermutationZscore}
\alias{PermutationZscore}
\title{We permute X}
\usage{
PermutationZscore(perm.nb)
}
\description{
for simplicity we assume that ncol(X) = 1
}
\details{
@export
}

