% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_phenotypeWay.R
\name{phenotypeWayReg_glm_score}
\alias{phenotypeWayReg_glm_score}
\title{lm + zscore in the phenotype way i.e X ~ G_j + U}
\usage{
phenotypeWayReg_glm_score(gif = FALSE, family = gaussian,
  factorized.X1 = FALSE)
}
\description{
we use R lm
}

