% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_phenotypeWay.R
\name{phenotypeWayReg_lm_score}
\alias{phenotypeWayReg_lm_score}
\title{lm + zscore in the phenotype way i.e X ~ G_j + U}
\usage{
phenotypeWayReg_lm_score(calibrate = FALSE)
}
\description{
we use R lm
}

