% Generated by roxygen2: do not edit by hand
% Please edit documentation in Plink/Plink_LD.R
\name{Plink_LD_clumping}
\alias{Plink_LD_clumping}
\title{LD clumping with plink}
\usage{
Plink_LD_clumping(prefix.in, clump.p1 = 1e-04, clump.p2 = 0.01,
  clump.r2 = 0.5, clump.kb = 250)
}
\description{
see : http://zzz.bwh.harvard.edu/plink/clump.shtml
see: https://neurogenetics.qimrberghofer.edu.au/iSECA/LD_clumping_tutorial.html
}

