% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_RidgeLFMMMethod.R
\name{predict.RidgeLFMMMethod}
\alias{predict.RidgeLFMMMethod}
\title{prediction function for ridge lfmm.}
\usage{
\method{predict}{RidgeLFMMMethod}(m, dat)
}
\description{
Assume that fit has been run.
}
\details{
@export
}

