% Generated by roxygen2: do not edit by hand
% Please edit documentation in Predict/Predict_row.R
\name{Predict_row}
\alias{Predict_row}
\title{use G_i[j.assumed] to predict G_i[j.unknow]}
\usage{
Predict_row(m, X_i, G_i, j.assumed, j.unknow)
}
\description{
use G_i[j.assumed] to predict G_i[j.unknow]
}

