% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_RandomXSampler.R
\name{RandomXSampler}
\alias{RandomXSampler}
\title{Sample random X variable.}
\usage{
RandomXSampler(sampler, d = 1, X.sampler = function(n, d) MASS::mvrnorm(n =
  n, Sigma = 1 * diag(1, d, d), mu = rep(0, d)))
}
\description{
Sample random X variable.
}

