% Generated by roxygen2: do not edit by hand
% Please edit documentation in Helpers/dataReaders.R
\name{read_vcf}
\alias{read_vcf}
\title{read vcf}
\usage{
read_vcf(f, maf.threshold = NULL, n_max = Inf, block.size = 1e+05)
}
\arguments{
\item{maf.threshold}{filter with this maf threshold}
}
\description{
read vcf
}

