% Generated by roxygen2: do not edit by hand
% Please edit documentation in Helpers/dataReaders.R
\name{read_vcf_files}
\alias{read_vcf_files}
\title{read vcf files}
\usage{
read_vcf_files(file.pattern, maf.threshold = NULL, save.rds = FALSE,
  block.size = 1e+05)
}
\arguments{
\item{maf.threshold}{filter with this maf threshold}
}
\description{
Rmk: it uses readr::read_delim
}

