% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_refractor.R
\name{refactor}
\alias{refactor}
\title{code found here: https://github.com/cozygene/refactor ... surement codé par un expert en R ^_^'}
\usage{
refactor(O, k, covarfile = NULL, t = 500, numcomp = NULL, stdth = 0.02,
  out = "refactor")
}
\description{
code found here: https://github.com/cozygene/refactor ... surement codé par un expert en R ^_^'
}

