% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_refractor.R
\name{refractorMethod}
\alias{refractorMethod}
\title{refractorMethod}
\usage{
refractorMethod(K, t,
  hypothesis.testing.method = phenotypeWayReg_glm_score(gif = FALSE),
  name = "refractor", nickname = NULL, verbose = FALSE)
}
\description{
refractorMethod
}

