% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_bootstrap.R
\name{residualBootstrapLFMMWithMissing}
\alias{residualBootstrapLFMMWithMissing}
\title{boostrap method described in 8.3 of http://www.stat.ncsu.edu/people/lu/courses/ST505/Ch8.pdf}
\usage{
residualBootstrapLFMMWithMissing(dat, m)
}
\description{
We sample a dataset with missing value at same place than in G
}

