% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_RidgeLFMMMethod.R
\name{RidgeLFMMMethod}
\alias{RidgeLFMMMethod}
\title{||G  - U V^t - X B || + lambda ||B||^2}
\usage{
RidgeLFMMMethod(K, hypothesis.testing.method = Zscore(NormalZscore(),
  AnalyticSigma2Functor()), lambda = 1, center = TRUE,
  name = "RidgeLFMMMethod", nickname = NULL, reuse.V = FALSE)
}
\description{
||G  - U V^t - X B || + lambda ||B||^2
}

