% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_RidgeLFMMMethod.R
\name{RidgeLFMMMethod_main}
\alias{RidgeLFMMMethod_main}
\title{Compute P, U, V, C and B of the ridge LFMM method}
\usage{
RidgeLFMMMethod_main(m, G_, dat, lambda, reuse)
}
\description{
Compute P, U, V, C and B of the ridge LFMM method
}

