% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_FromTrueData2.R
\name{sampl.FromTrueSampler2}
\alias{sampl.FromTrueSampler2}
\title{Sample data from true dataset version 2}
\usage{
\method{sampl}{FromTrueSampler2}(s)
}
\description{
Sample data from true dataset version 2
}

