% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_GenerativeSampler.R
\name{sampl.LogisticSampler}
\alias{sampl.LogisticSampler}
\title{Sample P(G) = logistic(mu + UV^T + XB^T + e)}
\usage{
\method{sampl}{LogisticSampler}(x)
}
\description{
Sample P(G) = logistic(mu + UV^T + XB^T + e)
}

