% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_GenerativeSampler.R
\name{sampl.NormalSampler2}
\alias{sampl.NormalSampler2}
\title{Sample G = mu + UV^T + XB^T + e}
\usage{
\method{sampl}{NormalSampler2}(s)
}
\description{
Sample G = mu + UV^T + XB^T + e
}

