% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_helpers.R
\name{sample_correlated_X}
\alias{sample_correlated_X}
\title{Sample a normaly distributed variable X such that cor(X,Y) = c}
\usage{
sample_correlated_X(Y, c, mu = 0, sd = 1)
}
\description{
Sample a normaly distributed variable X such that cor(X,Y) = c
}

