% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_helpers.R
\name{sample_not_null_rnorm}
\alias{sample_not_null_rnorm}
\title{Sample P(X| abs(X) > rho * sd ) ~ N(0,sd)}
\usage{
sample_not_null_rnorm(n, sd, rho)
}
\description{
Sample P(X| abs(X) > rho * sd ) ~ N(0,sd)
}

