% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_helpers.R
\name{sample_X_sum_correlated_U}
\alias{sample_X_sum_correlated_U}
\title{Sample X such that X = Sum(X_i) where i are corralated with Ui}
\usage{
sample_X_sum_correlated_U(U, cs)
}
\description{
Sample X such that X = Sum(X_i) where i are corralated with Ui
}

