% Generated by roxygen2: do not edit by hand
% Please edit documentation in Utils/utils.R
\name{save_plot_timc_bcm_15}
\alias{save_plot_timc_bcm_15}
\title{save plot to png format on timc-bcm-15}
\usage{
save_plot_timc_bcm_15(pl, filename, width, height)
}
\description{
save plot to png format on timc-bcm-15
}

