% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_MissingValueSampler.R
\name{SparseMissingValueSampler}
\alias{SparseMissingValueSampler}
\title{Sample missing values only on some loci}
\usage{
SparseMissingValueSampler(sampler, missing.prop, missing.prop.by.loci)
}
\description{
Sample missing values only on some loci
}

