% Generated by roxygen2: do not edit by hand
% Please edit documentation in DataSet/DataSet_TrueDataSet.R
\name{TrueDataSet}
\alias{TrueDataSet}
\title{TrueDataSet class}
\usage{
TrueDataSet(G = matrix(0, 1, 10), X = matrix(0, 2, 1), outlier = c(),
  reference = FALSE)
}
\description{
TrueDataSet class
}

