% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_helpers.R
\name{tscoreToPvalue}
\alias{tscoreToPvalue}
\title{score are assume to follow student distibution with df degre of freedom}
\usage{
tscoreToPvalue(score, df)
}
\description{
score are assume to follow student distibution with df degre of freedom
}

