% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod.R
\name{UXlink.Method}
\alias{UXlink.Method}
\title{average abs correlation between U and X}
\usage{
\method{UXlink}{Method}(m, dat, link.stat = function(x, y) abs(cor(x, y)),
  summary.stat = mean)
}
\description{
average abs correlation between U and X
}

