% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_MafFilter.R
\name{VarFilterSampler}
\alias{VarFilterSampler}
\title{run the sampler and filter column without variation}
\usage{
VarFilterSampler(sampler)
}
\description{
run the sampler and filter column without variation
}

