% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_zscore.R
\name{Zscore}
\alias{Zscore}
\title{normal_zscore functor}
\usage{
Zscore(zscorepvalue.functor = NormalZscore(),
  B.sigma2.functor = AnalyticSigma2Functor(), H0.mean = zero)
}
\description{
score = (B-0) / B.sigma2
}

