% Generated by roxygen2: do not edit by hand
% Please edit documentation in HypothesisTesting/HypothesisTesting_helpers.R
\name{zscoreToPvalue}
\alias{zscoreToPvalue}
\title{score are assume to follow normal distibution}
\usage{
zscoreToPvalue(score, mean = 0, sd = 1)
}
\description{
score are assume to follow normal distibution
}

