% Generated by roxygen2: do not edit by hand
% Please edit documentation in 3Article/3Article_MethodComparison.R
\name{Article3_MethodComparison}
\alias{Article3_MethodComparison}
\title{Article3_MethodComparison}
\usage{
Article3_MethodComparison(G.file, outlier.props = c(0.01, 0.05, 0.1),
  n = NULL, L = 10000, K = 4, K.method = K, correctionByC = FALSE,
  cs = c(0.2, 0.4, 0.6, 0.8), cs.sum = TRUE, nb.rep = 5,
  fast.only = TRUE, cluster.nb = NULL, save = TRUE, bypass = FALSE)
}
\arguments{
\item{outlier.props}{proportion of outlier simulated}

\item{cs}{if cs.sum = FALSE, correlation between PC1 and X. If cs.sum = TRUE, sum(PC,X) = cs}
}
\description{
Article3_MethodComparison
}

