% Generated by roxygen2: do not edit by hand
% Please edit documentation in Helpers/helpers.R
\name{DebugMessage}
\alias{DebugMessage}
\title{Print a message if getOption("Article3Package.debug") is not null}
\usage{
DebugMessage(msg)
}
\arguments{
\item{msg}{The message}
}
\description{
Print a message if getOption("Article3Package.debug") is not null
}

