% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_LassoLFMMMethod.R
\name{LassoLFMMMethod}
\alias{LassoLFMMMethod}
\title{||G  - U V^t - X B || + lambda |B|}
\usage{
LassoLFMMMethod(K, it.max, err.max, lambda = 0, lambda.K = 100,
  lambda.eps = 0.001, sparse.prop = NULL, center = TRUE,
  name = "LassoLFMMMethod", nickname = NULL,
  hypothesis.testing.method = NULL)
}
\description{
||G  - U V^t - X B || + lambda |B|
}

