% Generated by roxygen2: do not edit by hand
% Please edit documentation in LFMethod/LFMethod_LeaLFMMMethod.R
\name{LeaLFMMMethod}
\alias{LeaLFMMMethod}
\title{LEA package wrapper}
\usage{
LeaLFMMMethod(K, hypothesis.testing.method = NULL, iterations = 10000,
  CPU = 1, burnin = 5000, name = "LeaLFMMMethod", nickname = NULL,
  verbose = TRUE)
}
\description{
LEA package wrapper
}

