% Generated by roxygen2: do not edit by hand
% Please edit documentation in Sampler/Sampler_TrueSampler.R
\name{TrueSampler}
\alias{TrueSampler}
\title{sample true data from file}
\usage{
TrueSampler(G.file, X.file, outlier.file, n = NULL, L = NULL,
  reference = FALSE)
}
\arguments{
\item{G.file}{a file or a matrix}

\item{X.file}{a file or a matrix}

\item{outlier.file}{a file or a matrix}

\item{n}{number of indiv}

\item{L}{number of locus}

\item{reference}{if TRUE the sampler retrun R5 object}
}
\description{
sample true data from file
}

